Moved some code out of pipeline.rs
