+++ src/core/pipeline.rs
@@ -6,11 +6,6 @@ use std::{cell::RefCell, rc::Rc};
 use super::{Canvas, Instance};
 use crate::halw;
 
-pub trait Renderable {
-    fn stride() -> u32;
-    fn render(&self, command_buffer: &mut halw::CommandBuffer);
-}
-
 pub use hal::pso::{
     DescriptorArrayIndex, DescriptorBinding, DescriptorSetLayoutBinding, ShaderStageFlags,
     VertexBufferDesc,
@@ -181,25 +176,6 @@ impl From<hal::device::ShaderError> for PipelineCreationError {
 impl From<hal::pso::CreationError> for PipelineCreationError {
     fn from(e: hal::pso::CreationError) -> Self {
         PipelineCreationError::PipelineCreationFailed(e)
-    }
-}
-
-#[derive(Debug, PartialEq, Clone)]
-pub enum RenderingError {
-    NotProcessingFrame,
-}
-
-impl std::fmt::Display for RenderingError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            RenderingError::NotProcessingFrame => write!(f, "No frame is being processed"),
-        }
-    }
-}
-
-impl std::error::Error for RenderingError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        None
     }
 }
